refactor: file/folder structure
