+++ packages/client/src/pages/detail/[id].tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import DetailContent from "../../components/detail-content";
 import { useQuery } from "@apollo/client";
-import { GET_ITEM } from "../../controller/get-item";
+import { GET_ITEM } from "../../queries/get-item";
 import { navigate } from "gatsby";
 import "../../components/detail-rows/detail-rows.css";
 
@@ -15,8 +15,6 @@ const DetailPage = ({ params }: DetailPageProps): JSX.Element => {
 	const { data, error, loading } = useQuery(GET_ITEM, {
 		variables: { id: params.id },
 	});
-
-	console.log({ data });
 
 	if (error) {
 		navigate("/404", { replace: true });
